+++ src/Simplify.Web/Model/Validation/Attributes/MinAttribute.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Globalization;
 using System.Reflection;
 using Simplify.DI;
 
@@ -10,17 +11,74 @@ namespace Simplify.Web.Model.Validation.Attributes;
 /// <remarks>
 /// Initializes a new instance of the <see cref="MinAttribute"/> class.
 /// </remarks>
-/// <param name="minValue">Minimum value of the property.</param>
-/// <param name="errorMessage">The error message.</param>
-/// <param name="isMessageFromStringTable">if set to <c>true</c> [is message from string table].</param>
 [AttributeUsage(AttributeTargets.Property)]
-public class MinAttribute(IComparable minValue, string? errorMessage = null, bool isMessageFromStringTable = true)
-	: ValidationAttribute(errorMessage, isMessageFromStringTable)
+public class MinAttribute : ValidationAttribute
 {
 	/// <summary>
 	/// Gets the minimum value of the property.
 	/// </summary>
-	public IComparable MinValue { get; } = minValue;
+	/// <param name="minValue">Minimum value of the property.</param>
+	/// <param name="errorMessage">The error message.</param>
+	/// <param name="isMessageFromStringTable">if set to <c>true</c> [is message from string table].</param>
+	public MinAttribute(int minValue, string? errorMessage = null, bool isMessageFromStringTable = true) : base(errorMessage,
+		isMessageFromStringTable)
+	{
+		MinValue = minValue;
+		OperandType = typeof(int);
+	}
+
+	/// <summary>
+	/// Initializes a new instance of the <see cref="MinAttribute"/> class.
+	/// </summary>
+	/// <param name="minValue">Minimum value of the property.</param>
+	/// <param name="errorMessage">The error message.</param>
+	/// <param name="isMessageFromStringTable">if set to <c>true</c> [is message from string table].</param>
+	public MinAttribute(long minValue, string? errorMessage = null, bool isMessageFromStringTable = true) : base(errorMessage,
+		isMessageFromStringTable)
+	{
+		MinValue = minValue;
+		OperandType = typeof(long);
+	}
+
+	/// <summary>
+	/// Initializes a new instance of the <see cref="MinAttribute"/> class.
+	/// </summary>
+	/// <param name="minValue">Minimum value of the property.</param>
+	/// <param name="errorMessage">The error message.</param>
+	/// <param name="isMessageFromStringTable">if set to <c>true</c> [is message from string table].</param>
+	public MinAttribute(double minValue, string? errorMessage = null, bool isMessageFromStringTable = true) : base(errorMessage,
+		isMessageFromStringTable)
+	{
+		MinValue = minValue;
+		OperandType = typeof(double);
+	}
+
+	/// <summary>
+	/// Initializes a new instance of the <see cref="MinAttribute"/> class.
+	/// </summary>
+	/// <param name="type">The type of the minimum value of the property.</param>
+	/// <param name="minValue">Minimum value of the property.</param>
+	/// <param name="errorMessage">The error message.</param>
+	/// <param name="isMessageFromStringTable">if set to <c>true</c> [is message from string table].</param>
+	public MinAttribute(Type type, string minValue, string? errorMessage = null, bool isMessageFromStringTable = true) : base(errorMessage,
+		isMessageFromStringTable)
+	{
+		MinValue = minValue;
+		OperandType = type;
+	}
+
+	/// <summary>
+	/// Gets or sets the minimum value of the property.
+	/// </summary>
+	/// <value>
+	/// The minimum value of the property.
+	/// </value>
+	public object MinValue { get; }
+
+	/// <summary>
+	/// Gets the type of the minimum value.
+	/// </summary>
+	public Type OperandType { get; }
 
 	/// <summary>
 	/// Validates the specified property value.
@@ -33,21 +91,37 @@ public class MinAttribute(IComparable minValue, string? errorMessage = null, boo
 		if (value == null)
 			return;
 
-		if (value is not IComparable comparableValue)
-			throw new ArgumentException($"The type of specified property value must be inherited from {typeof(IComparable)}");
+		var minValue = ConvertToOperandComparableType(MinValue);
 
-		ValidateTypesMatching(comparableValue);
+		var comparableValue = ConvertToIComparable(value);
+
+		ValidateTypesMatching(value);
 
 		TryThrowCustomOrStringTableException(resolver);
 
-		if (comparableValue.CompareTo(MinValue) < 0)
+		if (comparableValue.CompareTo(minValue) < 0)
 			throw new ModelValidationException(
 				$"Property '{propertyInfo.Name}' required minimum value is {MinValue}, actual value: {value}");
 	}
 
-	private void ValidateTypesMatching(IComparable comparableValue)
+	private void ValidateTypesMatching(object value)
 	{
-		if (comparableValue.GetType() != MinValue.GetType())
+		if (value.GetType() != OperandType)
 			throw new ArgumentException("Type mismatch. The minimum value and property value should be of the same type.");
+	}
+
+	private IComparable ConvertToOperandComparableType(object value)
+	{
+		var convertedValue = Convert.ChangeType(value!, OperandType, CultureInfo.InvariantCulture);
+
+		return ConvertToIComparable(convertedValue);
+	}
+
+	private IComparable ConvertToIComparable(object value)
+	{
+		if (value is not IComparable comparableValue)
+			throw new ArgumentException($"The type of object value must be inherited from {typeof(IComparable)}");
+
+		return comparableValue;
 	}
 }