+++ src/Simplify.Web/CHANGELOG.md
@@ -32,6 +32,13 @@
 - Simplify.Templates bump to 2.0.2
 - Internal Simplify.Xml.Sources bump to 1.4
 
+## [4.9.1] - 2024-04-16
+
+### Fixed
+
+- Model Validation Attribute For Range Of Numbers (#257)
+- Model Validation Attributes For Min Max (#256)
+
 ## [4.9.0] - 2024-01-08
 
 ### Added